Fix a race condition
Fully reconsider the way send() synchronizes with receive(). Get rid of
"on data" object approach which proved to be non-atomic. Instead
properly orchestrate atomic updates of $!elems and 'receive awaiting'
promise.

And we're not truly lockless anymore because optimization of receive()
in concurrent threads requires them to synchronize at some point.